requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.4
httpx==0.25.1
python-multipart==0.0.6
PyJWT==2.8.0